+++ mlflow/transformers/flavor_config.py
@@ -8,9 +8,9 @@ from packaging.version import Version
 
 from mlflow.exceptions import MlflowException
 from mlflow.protos.databricks_pb2 import ALREADY_EXISTS, INVALID_PARAMETER_VALUE
-from mlflow.transformers.hub_utils import get_latest_commit_for_repo
 from mlflow.transformers.peft import _PEFT_ADAPTOR_DIR_NAME, get_peft_base_model, is_peft_model
 from mlflow.transformers.torch_utils import _extract_torch_dtype_if_set
+from mlflow.utils.huggingface_utils import get_latest_commit_for_repo
 
 if TYPE_CHECKING:
     import transformers
