--- mlflow/transformers/hub_utils.py
@@ -1,4 +1,0 @@
-# Re-export from the shared location for backward compatibility.
-from mlflow.utils.huggingface_utils import get_latest_commit_for_repo, is_valid_hf_repo_id
-
-__all__ = ["get_latest_commit_for_repo", "is_valid_hf_repo_id"]
